test: test gql query currently failing with cryptic error
