+++ packages/backend/src/express.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import server from "./apolloServer"
+
+const app = express();
+
+server.applyMiddleware({ app });
+
+export { app as default };
