+++ packages/backend/tests/utils/getClient.ts
@@ -1,0 +1,9 @@
+import ApolloBoost from 'apollo-boost';
+
+const getClient = () => {
+    return new ApolloBoost({
+        uri: `http://localhost:4000`
+    });
+};
+
+export { getClient }
