+++ packages/backend/tests/utils/globalSetup.ts
@@ -1,6 +1,10 @@
 require('ts-node/register');
 
+const server = require('../../src/express').default;
+const port = process.env.PORT || 4000;
+
 module.exports = async () => {
-  // code to run before entire testsuite, like starting the server
+  // tslint:disable-next-line
+  global['httpServer'] = await server.listen(port);
   return null;
 };
